+++ lib/screens/transaction_screens/update_transaction_screen.dart
@@ -60,7 +60,7 @@ class _UpdateTransactionScreenState extends State<UpdateTransactionScreen> {
   void getBalance() async {
     DataSnapshot snapshot = await dbBalance.child("user0").get();
     Map balance = snapshot.value as Map;
-    _amount = balance['amount'];
+    _amount = balance['amount'].toDouble();
   }
 
   @override
